+++ common/src/main/java/net/bettercombat/client/animation/PlayerAttackAnimatable.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface PlayerAttackAnimatable {
     void updateAnimationsOnTick();
     void playAttackAnimation(String name, AnimatedHand hand, float length, float upswing);
-    void stopAttackAnimation();
+    void stopAttackAnimation(float length);
 }