+++ common/src/main/java/net/bettercombat/mixin/client/AbstractClientPlayerEntityMixin.java
@@ -283,11 +283,13 @@ public abstract class AbstractClientPlayerEntityMixin extends PlayerEntity imple
     // PlayerAttackAnimatable
 
     @Override
-    public void stopAttackAnimation() {
+    public void stopAttackAnimation(float length) {
         IAnimation currentAnimation = attackAnimation.base.getAnimation();
         if (currentAnimation != null && currentAnimation instanceof KeyframeAnimationPlayer) {
+            var fadeOut = Math.round(length);
+            // attackAnimation.adjustmentModifier.fadeOut(fadeOut);
             attackAnimation.base.replaceAnimationWithFade(
-                    AbstractFadeModifier.standardFadeIn(5, Ease.INOUTSINE), null);
+                    AbstractFadeModifier.standardFadeIn(fadeOut, Ease.INOUTSINE), null);
         }
     }