+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add attack speed cap
 - Improve back to back animation overlapping
 - Fix offhand arm invisible while 2H animations
+- Rework feinting
 
 # 1.5.4
 - Fix attacking hostile mounts/vehicles (Thanks to Emily Darrow)